### 2020-06-15 #2240 @arfon
> @whedon invite @KedoKudo as reviewer

@jgostick - Whedon doesn't know how to invite reviewers like this yet sorry. If you want to invite @KedoKudo as a reviewer you can either do that in an email (e.g. [like this](https://joss.readthedocs.io/en/latest/editing.html#sample-email-to-potential-reviewers)) or just by mentioning them here.
