### 2020-06-15 #2240 @jgostick
This email template is what I needed.  Will send asap.  It'd be sweet if Whedon could send this though!
