### 2020-06-23 #2240 @arfon
> Hi @KedoKudo, do you need any help getting going on your review? I'm still looking for a second reviewer to add, but you can get started.

@jgostick - if you'd like to start the review with one reviewer (which is fine as long is there is a second one later), you should probably move this forward to the review stage where @KedoKudo will have a checklist and guidelines to work through. You can do this with `@whedon start review`
