### 2020-06-23 #2381 @whedon
[REVIEW]: xrdfit: A Python package for fitting synchrotron X-ray diffraction spectra
**Submitting author:** @merrygoat (<a href="http://orcid.org/0000-0002-5430-6924">Peter Crowther</a>)
**Repository:** <a href="https://github.com/LightForm-group/xrdfit" target ="_blank">https://github.com/LightForm-group/xrdfit</a>
**Version:** v1.1.2
