### 2020-07-02 7f9a60d
Added log scale option to plots
Relates to #35.
