+++ xrdfit/spectrum_fitting.py
@@ -166,17 +166,19 @@ class PeakFit:
         self._maxima_snrs: List[float] = []
 
     def plot(self, time_step: str = None, file_name: str = None, title: str = None,
-             label_angle: float = None):
+             label_angle: float = None, log_scale=False):
         """ Plot the raw spectral data and the fit.
         :param time_step: If provided, a time step used to generate the title of the plot.
         :param file_name: If provided, save the plot to this file as well as displaying it.
         :param title: If provided, override the autogenerated plot title with this title.
         :param label_angle: The angle to rotate maxima labels.
+        :param log_scale: Whether to plot the y-axis in a log or linear scale.
         """
         if self.raw_spectrum is None:
             print("Cannot plot fit peak as fitting has not been done yet.")
         else:
-            plotting.plot_peak_fit(self, time_step, file_name, title, label_angle)
+            plotting.plot_peak_fit(self, time_step, file_name, title, label_angle,
+                                   log_scale=log_scale)
 
     def get_maxima_snrs(self) -> List[float]:
         """Get the signal to noise ratio for each maxima in a PeakFit."""
@@ -245,7 +247,7 @@ class FitSpectrum:
         plotting.plot_polar_heat_map(num_cakes, rad, z_data, self.first_cake_angle)
 
     def plot(self, cakes_to_plot: Union[int, List[int]], x_range: Tuple[float, float] = None,
-             merge_cakes: bool = False, show_points=False):
+             merge_cakes: bool = False, show_points=False, log_scale=False):
         """Plot the intensity as a function of two theta for a given cake.
 
         :param cakes_to_plot: The numbers of one or more cakes to plot.
@@ -253,6 +255,7 @@ class FitSpectrum:
         :param merge_cakes: If True plot the sum of the selected cakes as a single line. If False
           plot all selected cakes individually.
         :param show_points: Whether to show data points on the plot.
+        :param log_scale: Whether to plot the y-axis on a log or linear scale.
         """
         if isinstance(cakes_to_plot, int):
             cakes_to_plot = [cakes_to_plot]
@@ -261,11 +264,11 @@ class FitSpectrum:
             data = self._get_spectrum_subset(cakes_to_plot, x_range, True)
         else:
             data = self.spectral_data
-        plotting.plot_spectrum(data, cakes_to_plot, merge_cakes, show_points, x_range)
+        plotting.plot_spectrum(data, cakes_to_plot, merge_cakes, show_points, x_range, log_scale)
         plt.show()
 
     def plot_fit(self, fit_name: str, time_step: str = None, file_name: str = None,
-                 label_angle: float = None):
+                 label_angle: float = None, log_scale=False):
         """Plot the result of a fit and the raw data.
 
         :param fit_name: The name of the fit to plot.
@@ -273,14 +276,15 @@ class FitSpectrum:
         :param file_name: If provided, the stub of the file name to write the plot to, if not
           provided, the plot will be displayed on screen.
         :param label_angle: The angle to rotate maxima labels.
+        :param log_scale: Whether to plot the y-axis on a log or linear scale.
         """
         fit = self.get_fit(fit_name)
-        fit.plot(time_step, file_name, label_angle=label_angle)
+        fit.plot(time_step, file_name, label_angle=label_angle, log_scale=log_scale)
 
     def plot_peak_params(self, peak_params: Union[PeakParams, List[PeakParams]],
                          cakes_to_plot: Union[int, List[int]],
                          x_range: Tuple[float, float] = None, merge_cakes: bool = False,
-                         show_points=False, label_angle=None):
+                         show_points=False, label_angle=None, log_scale=False):
         """Plot a visualisation of the provided :class:`PeakParams` over the raw data.
 
[…]
