+++ xrdfit/plotting.py
@@ -52,7 +52,7 @@ def plot_polar_heat_map(num_cakes: int, rad: List[int], z_data: np.ndarray, firs
 
 
 def plot_spectrum(data: np.ndarray, cakes_to_plot: List[int], merge_cakes: bool, show_points: bool,
-                  x_range: Union[None, Tuple[float, float]] = None):
+                  x_range: Union[None, Tuple[float, float]] = None, log_scale=False):
     """Plot a raw spectrum using matplotlib.
 
     :param data: The data to plot, x_data in column 0, y data in columns 1-N where N is the number
@@ -62,6 +62,7 @@ def plot_spectrum(data: np.ndarray, cakes_to_plot: List[int], merge_cakes: bool,
       all selected cakes individually.
     :param show_points: Whether to show data points on the plot.
     :param x_range: If supplied, restricts the x-axis of the plot to this range.
+    :param log_scale: If True, plot y axis on log scale. If False use linear scale.
     """
     if show_points:
         line_spec = "-x"
@@ -86,6 +87,8 @@ def plot_spectrum(data: np.ndarray, cakes_to_plot: List[int], merge_cakes: bool,
     plt.ylabel('Intensity')
     if x_range:
         plt.xlim(x_range[0], x_range[1])
+    if log_scale:
+        plt.yscale("log")
     plt.tight_layout()
 
 
@@ -120,7 +123,7 @@ def plot_peak_params(peak_params: List["PeakParams"], x_range: Tuple[float, floa
 
 
 def plot_peak_fit(peak_fit: "PeakFit", time_step: str = None, file_name: str = None,
-                  title: str = None, label_angle: float = None):
+                  title: str = None, label_angle: float = None, log_scale=False):
     """Plot the result of a peak fit as well as the raw data.
 
     :param peak_fit: The result of a peak fit
@@ -128,6 +131,7 @@ def plot_peak_fit(peak_fit: "PeakFit", time_step: str = None, file_name: str = N
     :param file_name: If provided used as a on disk location to save the plot.
     :param title: If provided, can be used to override the auto generated plot title.
     :param label_angle: The angle to rotate maxima labels.
+    :param log_scale: Whether to plot the y axis on a log or linear scale.
     """
     data = peak_fit.raw_spectrum
     # First plot the raw data
@@ -155,7 +159,8 @@ def plot_peak_fit(peak_fit: "PeakFit", time_step: str = None, file_name: str = N
         maxima_center = peak_fit.result.params[f"maximum_{index}_center"]
         plt.text(maxima_center, plt.ylim()[1] * 1.05, maxima_name, horizontalalignment="center",
                  fontsize=title_size * 0.8, rotation=label_angle)
-
+    if log_scale:
+        plt.yscale("log")
     plt.tight_layout()
     if file_name:
         file_name = pathlib.Path(file_name)
@@ -168,7 +173,7 @@ def plot_peak_fit(peak_fit: "PeakFit", time_step: str = None, file_name: str = N
 
 
 def plot_parameter(data: np.ndarray, fit_parameter: str, show_points: bool,
-                   show_error: bool, scale_by_error: bool = False):
+                   show_error: bool, scale_by_error: bool = False, log_scale=False):
     """Plot a parameter of a fit against time.
 
     :param data: The data to plot, x data in the first column, y data in the second column and
@@ -178,6 +183,7 @@ def plot_parameter(data: np.ndarray, fit_parameter: str, show_points: bool,
     :param show_error: Whether to show error bars on the plot.
     :param scale_by_error: If True auto scale the y-axis to the range of the error bars. If False,
       auto scale the y-axis to the range of the data.
[…]
