### 2020-07-02 34e1f1d
Rotate maps according to first cake angle
Fixes #35.
