### 2020-07-08 LightForm-group/xrdfit#38 @KedoKudo
[JOSS REVIEW] Several questions 
Hi @merrygoat and @christopher-daniel,

I have several questions regarding your package `xrdfit`, and I think addressing them can help it reach a larger audience in the long run

* Benchmark comparison
Have you performed any benchmark comparison for `xrdfit` with other similar software(s) in the field?  It would be helpful to let users know its performance and (potentially) scalability so that they can decide whether it is suitable for their data.

* JOSS paper related
    * The DOI listed in the paper does not exist (typo perhaps?)
    * Once the preprint (Direct Evidence for a Dynamic Phase Transformation during High Temperature Deformation in Ti-64) published, please update the reference section accordingly

* Distributed computing
So far it seems like `xrdfit` is a single node application.  Will it provide parallel capabilities (MPI) in the future?
