### 2020-07-08 LightForm-group/xrdfit#39 @KedoKudo
[JOSS REVIEW] Enhancement suggestions
Hi @merrygoat and @christopher-daniel,

After testing `xrdfit` locally, I would like to make the following enhancement suggestions:

- [x]  Put the repo on Binder (https://mybinder.org) or Colab (colab.research.google.com) so that others can directly test them in the browser without installing

    * please use either `pip freeze` or `conda env export` to generate `requirement.txt` so that it contains necessary version information
    * according to my test, `xrdfit` also needs `ipywidgets` to function properly in JupyterLab, otherwise an error will occur in the 3rd example notebook.

- [x]  add unit testing
please consider adding testing (unit testing) to `xrdfit`.

- [x] add community guideline
It would be helpful to have a community guideline (such as code of conduct, how to contribute to the project) for `xrdfit`.
