### 2020-07-08 #2381 @KedoKudo
@jgostick - I finished my review of `xrdfit` and submitted my feedbacks to the developers on the corresponding repo (https://github.com/LightForm-group/xrdfit/issues/39, https://github.com/LightForm-group/xrdfit/issues/38).

@merrygoat and @christopher-daniel - please feel free to ping me if you have any questions about the feedbacks.
