### 2020-07-08 #2381 @merrygoat
@KedoKudo Thank you very much for your feedback. I will address the feedback items individually on the issues you have raised in the `xrdfit` repository.
