### 2020-07-08 29deb14
Freeze package versions and add ipywidgets
Fixes #41.
