+++ requirements.txt
@@ -1,7 +1,8 @@
-numpy
-matplotlib
-pandas
-dill
-tqdm
-lmfit
-notebook
+numpy==1.19.0
+matplotlib==3.2.2
+pandas==1.0.5
+dill==0.3.2
+tqdm==4.47.0
+lmfit==1.0.1
+notebook==6.0.3
+ipywidgets==7.5.1
