### 2020-07-08 LightForm-group/xrdfit#38 @merrygoat
Thank you for your comments.

- We haven't made direct comparisons to other software. As I understand, similar software (MAUD being one example) tends to base fits on an underlying model of the scattering mechanism whereas `xrdfit` takes the simpler approach of fitting the raw data without considering the underlying model of scattering. I agree that it would be a good idea to add some approximate information about performance and scaling to the documentation and have opened issue #43. Perhaps @christopher-daniel could comment on other software in the field and their approach?

- I think all of the DOIs in the paper work correctly. Could you provide more information on which one is broken? We will update the reference to `Direct Evidence for a Dynamic Phase Transformation` if it is published before `xrdfit`.

- Yes, you are correct, it is single core. I have found implementing multiple processing difficult (the weakness of Python!).

The fitting of multiple time steps cannot be done in parallel, since the start point of each fit depends on the last step and so they must be carried out in order.

The fitting of the different peaks within a single pattern should be possible to run over multiple cores and I consider this in issue #28 (test code is in the multiprocessing branch). The problem is that this makes performance slower than running single core, I think because each fit is modest (~ 2 seconds), the overhead is greater than the computation. I will continue investigating a little more as it may be a weakness in my implementation, though I don't see anything obvious.

For performance in general. profiling shows 95% of the performance time in `xrdfit` is taken by lmfit. lmfit is a layer to the scipy.optimize.curve_fit function and so has some overhead. It should be possible to directly implement the fitting using SciPy to remove this overhead. However, profiling shows that the overhead of lmfit in this case is not large - perhaps 20-30% overall, so I  do not think that this change would be transformative. Since lmfit provides many convenience functions, greatly reducing the complexity of the code, I think that the performance cost is worth it.

Using something like MPI would require stepping outside Python to write fitting functions and I think that is beyond the scope of this project (and beyond my abilities!).
