### 2020-07-09 LightForm-group/xrdfit#38 @christopher-daniel
It is possible to run single peak fitting in [DAWN](https://dawnsci.org), an in-house software developed at Diamond Light Source. This is briefly mentioned in the paper introduction, as well as mentioned in the notebooks describing the initial 'caking' of the data. There is also [Polydefix](http://merkel.texture.rocks/Multifit-Polydefix/?lang=en), although I am less familiar with this software. But, in both cases, the software is a compiled package and uses a GUI. So, the peak fitting in DAWN involves a point and click method to select the peak bounds, meaning it isn't necessarily a reproducible analysis. This does an ok job for small datasets with clearly defined peaks, but would clearly take a long time for any systematic analysis of many different tests or cakes. However, and most importantly, it is difficult to clearly distinguish any multiple peaks with any crossover using DAWN, with the peak fits either merging into one, or an error message appearing asking the user to redefine the peak bounds each time. This really makes any analysis fitting multipeaks in datasets containing 1000s of patterns impossible with these current available software.

MAUD is a different software which does a different job. This uses a Rietveld refinement method to match a model of the material properties (and the beamline setup) to the data. This is great for determining things like the phase fraction and crystal orientation, etc. But, because MAUD uses a model of the unit cell to fit the data, there is some averaging of the overall fit and a larger error in the individual peak position. So, we really need xrdfit for an accurate single peak fitting of the individual lattice plane peaks, to determine their peak position and the peak shifts over time, which is used to calculate the micromechanical behaviour.

I will run some quick (rough) benchmark tests for a single peak fit in DAWN and compare this to xrdfit.
