### 2020-07-09 LightForm-group/xrdfit#38 @KedoKudo
@merrygoat and @christopher-daniel,  thanks for the detailed explanation.  I am looking forward to the benchmark results 😄
