### 2020-07-16 #2381 @jgostick
Thanks @KedoKudo! @merrygoat can you let me know in this issue when you've made the changes?

@mikapfl, how is your review coming along?  It's only been 3 weeks, but things seem to move so fast at JOSS.
