### 2020-07-22 03ba476
Add info about other packages to paper
