+++ paper/paper.md
@@ -26,9 +26,11 @@ bibliography: paper.bib
 
 # Summary
 
-The evolution of peak profiles in synchrotron X-ray diffraction (SXRD) data can tell us how the internal crystallographic structures of metals change during applied heating, high temperature straining and cooling cycles [@Daniel_2019; @Stark_2015; @Canelo_Yubero_2016; @Hu_2017], which is invaluable information used to improve industrial processing routes [@Salem_2008]. The experiment requires a beamline, such as Diamond Light Source [@Diamond_2020], to produce a high energy X-ray beam and illuminate a polycrystalline sample [@Daniel_2019]. The results are recorded in the form of time-resolved diffraction pattern rings, which are converted into a spectra of intensity peaks versus two-theta angle for a given direction [@Filik_2017; @Ashiotis_2015; @Hammersley_1996]. However, since many intensity profiles are collected during each experiment, with detectors recording at speeds of up to 250 Hz [@Diamond2_2020], fitting each of the individual lattice plane peaks can take a long time using current available software [@Basham_2015; @Merkel_2015]. Distinguishing the evolution of any overlapping peaks, for multi-phase materials, can also be difficult and time-consuming. Therefore, a faster and more robust Python package has been developed to fit the evolution of multiple and overlapping peaks for datasets containing many thousands of SXRD patterns.
+The evolution of peak profiles in synchrotron X-ray diffraction (SXRD) data can tell us how the internal crystallographic structures of metals change during applied heating, high temperature straining and cooling cycles [@Daniel_2019; @Stark_2015; @Canelo_Yubero_2016; @Hu_2017], which is invaluable information used to improve industrial processing routes [@Salem_2008]. The experiment requires a beamline, such as Diamond Light Source [@Diamond_2020], to produce a high energy X-ray beam and illuminate a polycrystalline sample [@Daniel_2019]. The results are recorded in the form of time-resolved diffraction pattern rings, which are converted into a spectra of intensity peaks versus two-theta angle for a given direction [@Filik_2017; @Ashiotis_2015; @Hammersley_1996]. However, since many intensity profiles are collected during each experiment, with detectors recording at speeds of up to 250 Hz [@Diamond2_2020], fitting each of the individual lattice plane peaks can take a long time using current available software [@Basham_2015; @Merkel_2015].
 
-``xrdfit`` is a Python package for fitting the diffraction peaks in SXRD (and XRD) spectra. It is intended as an easy to use tool for the quick analysis of individual and overlapping lattice plane peaks, to discern the peak position and profile. The features of  ``xrdfit`` are shown schematically in \autoref{fig:figure1}. ``xrdfit`` uses the Python package ``lmfit`` [@Newville_2014] for the underlying fitting. Features are included for selecting different 'cakes' of data and automating fitting over many spectra, to enable tracking of peaks as they shift throughout the experiment. By analysing how different lattice plane peaks change during simulated processing, as can be seen in \autoref{fig:figure2}, the transformation and micromechanical behaviour of the material can be understood.
+There are existing packages which can be used to fit peaks in SXRD spectra, examples include DAWN [citation] and Mulifit [citation]. In both cases, the software are compiled packages which use a graphical interface. Setting up the peak fits involves a point and click method to select peak bounds, meaning it is hard to create a repeatable analysis. These packages also struggle in cases where peaks overlap. MAUD [citation] is a software pacakge that approaches fitting in a different way. MAUD uses the Rietveld refinement method to match a model of the material properties and the beamline setup to the data. This is method allows determination of some additional material properties but the method results in some error in individual peak positions meaning peak shifts cannot be accurately determined.
+
[…]
