### 2020-07-23 0e65060
Improve __str__ and __repr__ for objects
