### 2020-07-23 26ec676
Add logging for debug and warnings
Fixes #44.
