### 2020-07-23 737b901
Add comparison to other tools to documentation
