### 2020-07-23 edce593
Create CODE_OF_CONDUCT.md
