### 2020-07-23 LightForm-group/xrdfit#39 @merrygoat
I have added CODE_OF_CONDUCT.md along with the existing CONTRIBUTING.md

I don't think I have the knowledge to usefully add unit testing. I could test simple things like function arguments types and return types, though I feel this is mostly covered by the function type hints. A lot of the the functionality of the methods is then either plotting or fitting, both of which I don't know how I would test at a function level since there is no real "right" answer. A good fit is, to some extent, the one that looks good. I would be pleased for suggestions about a good way to do this though.

I treat the tutorial notebooks as informal integration tests, since they use all aspects of the program and we run through them each time we make a release to make sure everything is behaving as expected.
