### 2020-07-27 LightForm-group/xrdfit#39 @KedoKudo
@merrygoat Thanks for implementing the suggestions.  I'm closing this issue now 😄
