+++ .github/workflows/test_notebooks.yaml
@@ -1,0 +1,13 @@
+on: push
+jobs:
+  run:
+    runs-on: ubuntu-latest
+    name: Run treebeard
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+      - uses: treebeardtech/treebeard@master
+        notebooks: 
+          - "tutorial notebooks/1*"
+          - "tutorial notebooks/2*"
+          - "tutorial notebooks/3*"
