+++ .github/workflows/test_notebooks.yaml
@@ -7,7 +7,8 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
       - uses: treebeardtech/treebeard@master
-        notebooks: 
-          - "tutorial notebooks/1*"
-          - "tutorial notebooks/2*"
-          - "tutorial notebooks/3*"
+        with:
+          notebooks: 
+            - "tutorial notebooks/1*"
+            - "tutorial notebooks/2*"
+            - "tutorial notebooks/3*"
