+++ .github/workflows/test_notebooks.yaml
@@ -8,4 +8,4 @@ jobs:
       - uses: actions/setup-python@v2
       - uses: treebeardtech/treebeard@master
         with:
-          notebooks: "tutorial notebooks/1*"
+          notebooks: "tutorial notebooks/1*", "tutorial notebooks/2*", "tutorial notebooks/3*"
