### 2020-07-29 434660e
Updates to paper and references
Changes to main text.
Include new references to DAWN, Multifit/Polydefix, Fit2d, Maud, etc.
