+++ paper/paper.bib
@@ -156,3 +156,55 @@
   url       = {http://www.tandfonline.com/doi/abs/10.1080/08957959608201408},
 }
 
+@article{Hammersley2016,
+author = "Hammersley, A. P.",
+title = "{{\it FIT2D}: a multi-purpose data reduction, analysis and visualization program}",
+journal = "Journal of Applied Crystallography",
+year = "2016",
+volume = "49",
+number = "2",
+pages = "646--652",
+month = "Apr",
+doi = {10.1107/S1600576716000455},
+url = {https://doi.org/10.1107/S1600576716000455},
+abstract = {{\it FIT2D}~is one of the principal area detector data reduction, analysis and visualization programs used at the European Synchrotron Radiation Facility and is also used by more than 400 research groups worldwide, including many other synchrotron radiation facilities. It has been developed for X-ray science, but is applicable to other structural techniques and is used in analysing electron diffraction data and microscopy, and neutron diffraction and scattering data. {\it FIT2D}~works for both interactive and `batch'-style data processing. Calibration and correction of detector distortions, integration of two-dimensional data to a variety of one-dimensional scans, and one- and two-dimensional model fitting are the main uses. Many other general-purpose image processing and image visualization operations are available. Commands are available through a `graphical user interface' and operations common to certain types of analysis are grouped within `interfaces'. Executable versions for most workstation and personal computer systems, and web page documentation, are available at http://www.esrf.eu/computing/scientific/FIT2D.},
+keywords = {data analysis, data display, data reduction, data visualization, detector calibration, image processing, model fitting, powder diffraction, small-angle scattering, wide-angle scattering},
+}
+
+@article{Rietveld1969,
+abstract = {A structure refinement method is described which does not use integrated neutron powder intensities, single or overlapping, but employs directly the profile intensities obtained from step-scanning measurements of the powder diagram. Nuclear as well as magnetic structures can be refined, the latter only when their magnetic unit cell is equal to, or a multiple of, the nuclear cell. The least-squares refinement procedure allows, with a simple code, the introduction of linear or quadratic constraints between the parameters.},
+archivePrefix = {arXiv},
+arxivId = {10.1107/S0021889869006558},
+author = {Rietveld, H. M.},
+doi = {10.1107/S0021889869006558},
+eprint = {S0021889869006558},
+file = {:Users/mbcx9cd4/Documents/Mendeley Desktop/Rietveld - 1969 - A profile refinement method for nuclear and magnetic structures.pdf:pdf},
+isbn = {0021-8898},
+issn = {00218898},
+journal = {Journal of Applied Crystallography},
+month = {jun},
+number = {2},
+pages = {65--71},
+primaryClass = {10.1107},
+publisher = {International Union of Crystallography},
+title = {{A profile refinement method for nuclear and magnetic structures}},
+url = {http://scripts.iucr.org/cgi-bin/paper?S0021889869006558},
+volume = {2},
+year = {1969}
+}
+
+@article{Lutterotti2014,
+author = {Lutterotti, Luca and Vasin, Roman and Wenk, Hans-Rudolf},
+doi = {10.1017/S0885715613001346},
+file = {:Users/mbcx9cd4/Documents/Mendeley Desktop/Lutterotti, Vasin, Wenk - 2014 - Rietveld texture analysis from synchrotron diffraction images. I. Calibration and basic analysis.pdf:pdf},
+issn = {0885-7156},
+journal = {Powder Diffraction},
+keywords = {diffraction images,rietveld method,synchrotron diffraction,texture analysis},
+month = {mar},
+number = {01},
+pages = {76--84},
[…]
