### 2020-07-29 9f92c63
Add treebeard test badge to readme
