### 2020-07-29 854d69e
Update documentation file index.rst
Edit DAWN and MAUD comparison.
