### 2020-07-31 #2381 @mikapfl
Hi,

I've looked into the paper itself. I think it is a nice and short description of the the capabilities of `xrdfit` and motivates well what the purpose of `xrdfit` is. I found only a couple of minor things which could be clearer:
* In the first paragraph, you write:
> The experiment requires a beamline, such as Diamond Light Source […]

However, the Diamond Light Source is a synchrotron radiation facility comprising many beamlines, so this is not quite correct. Maybe something along the lines "The experiment requires a beamline at a synchrotron radiation facily such as […]" or such might be better.

* A little later you write:
> However, since many intensity profiles are collected during each
experiment, with detectors recording at speeds of up to 250 Hz (Diamond Light Source Ltd,
2020b),

Here, I think the web page of the I12 beamline at Diamond might not be an ideal citation. Maybe https://doi.org/10.1109/NSSMIC.2012.6551180 would be the best citation for the recording speeds of PILATUS3 X CdTe detectors? This is the best description of the PILATUS3 ASIC I know of. They specify a maximum frame rate of 500 Hz, which is not realized in the commercial 2M detector, which is described in detail at the web page of dectris: https://www.dectris.com/products/pilatus3/pilatus3-x-cdte-for-synchrotron/pilatus3-x-cdte-2m/ .

* In figure 2, your choice of the colours red and green for the two right axes and the lines for temperature and deformation is not optimal, because it is not colourblind-friendly. Consider using colours with different grey levels or something like dotted or dashed lines.

Best Regards,

Mika
