### 2020-08-03 72f9231
Update paper, figure and references for review
Update references in paper.bib and change text in paper.md as recommended by reviewer.
Update figure2.eps colour scheme to be colourblind-friendly.
