+++ paper/paper.bib
@@ -80,13 +80,6 @@
   url             = {https://www.diamond.ac.uk},
 }
 
-@Misc{Diamond2_2020,
-  author          = {{Diamond Light Source Ltd}},
-  title           = {{Detectors at I12}},
-  year            = {2020},
-  url             = {https://www.diamond.ac.uk/Instruments/Imaging-and-Microscopy/I12/Detectors-at-I12.html},
-}
-
 @Article{Filik_2017,
   author    = {J. Filik and A. W. Ashton and P. C. Y. Chang and P. A. Chater and S. J. Day and M. Drakopoulos and M. W. Gerring and M. L. Hart and O. V. Magdysyuk and S. Michalik and A. Smith and C. C. Tang and N. J. Terrill and M. T. Wharmby and H. Wilhelm},
   title     = {Processing two-dimensional X-ray diffraction and small-angle scattering data {inDAWN} 2},
@@ -156,7 +149,7 @@
   url       = {http://www.tandfonline.com/doi/abs/10.1080/08957959608201408},
 }
 
-@article{Hammersley2016,
+@article{Hammersley_2016,
 author = "Hammersley, A. P.",
 title = "{{\it FIT2D}: a multi-purpose data reduction, analysis and visualization program}",
 journal = "Journal of Applied Crystallography",
@@ -171,7 +164,7 @@ abstract = {{\it FIT2D}~is one of the principal area detector data reduction, an
 keywords = {data analysis, data display, data reduction, data visualization, detector calibration, image processing, model fitting, powder diffraction, small-angle scattering, wide-angle scattering},
 }
 
-@article{Rietveld1969,
+@article{Rietveld_1969,
 abstract = {A structure refinement method is described which does not use integrated neutron powder intensities, single or overlapping, but employs directly the profile intensities obtained from step-scanning measurements of the powder diagram. Nuclear as well as magnetic structures can be refined, the latter only when their magnetic unit cell is equal to, or a multiple of, the nuclear cell. The least-squares refinement procedure allows, with a simple code, the introduction of linear or quadratic constraints between the parameters.},
 archivePrefix = {arXiv},
 arxivId = {10.1107/S0021889869006558},
@@ -193,7 +186,7 @@ volume = {2},
 year = {1969}
 }
 
-@article{Lutterotti2014,
+@article{Lutterotti_2014,
 author = {Lutterotti, Luca and Vasin, Roman and Wenk, Hans-Rudolf},
 doi = {10.1017/S0885715613001346},
 file = {:Users/mbcx9cd4/Documents/Mendeley Desktop/Lutterotti, Vasin, Wenk - 2014 - Rietveld texture analysis from synchrotron diffraction images. I. Calibration and basic analysis.pdf:pdf},
@@ -208,3 +201,24 @@ url = {http://www.journals.cambridge.org/abstract{\_}S0885715613001346},
 volume = {29},
 year = {2014}
 }
+
+@Misc{PILATUS_2020,
+  author          = {{DECTRIS - detecting the future}},
+  title           = {{PILATUS3 X CdTe 2M}},
+  year            = {2020},
+  url             = {https://www.dectris.com/products/pilatus3/pilatus3-x-cdte-for-synchrotron/pilatus3-x-cdte-2m/},
+}
+
+@INPROCEEDINGS{Loeliger_2012,
+author={T. {Loeliger} and C. {Brönnimann} and T. {Donath} and M. {Schneebeli} and R. {Schnyder} and P. {Trüb}},
+booktitle={2012 IEEE Nuclear Science Symposium and Medical Imaging Conference Record (NSS/MIC)}, 
+title={The new PILATUS3 ASIC with instant retrigger capability}, 
+year={2012},
+volume={},
+number={},
+pages={610-615},
[…]
