### 2020-08-11 be7359e
Remove plt.close from plot_peak_fit
