### 2020-08-12 eee1ca3
Add delimiter option for reading data files.
