### 2020-08-12 f60c7bb
Increase gridsize by 1 to fix colormesh deprecation
