### 2020-08-12 c8a7272
Add option to load cakes anticlockwise
