### 2020-08-12 e3cda9e
Add new caking features to documentation
