### 2020-08-12 54d7048
Fix for cake highlighting
