### 2020-08-12 #2381 @merrygoat
Thank you yes - we have added those paper changes [here](https://github.com/LightForm-group/xrdfit/commit/72f923164bd66fb82eb92ae42f49ace08bd20053) and [here](https://github.com/LightForm-group/xrdfit/commit/434660e240a52a48e1e18f10e2fde18d184be9ed).

We are just finishing up the last of the issues the reviewers raised, and adding a little more documentation for the data preparation step and we should be good.
