### 2020-08-12 LightForm-group/xrdfit#38 @merrygoat
Sorry, forgot to update this. More information on other packages was added in 03ba476242599d3b5c97468c8d61213d917966fb and some information on performance in 854d69eb9a204db4f8082273a2dc50657a6a8d1c.

The good news is that xrdfit is quicker both in speed of analysis (very approximately 20 times quicker than DAWN)  but also quicker to set up the workflow to measure many peaks.
