+++ paper/paper.md
@@ -32,9 +32,9 @@ There are existing packages which can be used to fit peaks in SXRD spectra, exam
 
 ``xrdfit`` is a Python package we have developed for a faster fitting of diffraction peaks in SXRD (and XRD) spectra, which can be used for datasets containing many thousands of patterns. It is intended as an easy to use tool which enables automated, repeatable identification of peak positions and profiles in spectra with multiple individual or overlapping lattice plane peaks. The features of  ``xrdfit`` are shown schematically in \autoref{fig:figure1}. ``xrdfit`` uses the Python package ``lmfit`` [@Newville_2014] for the underlying fitting. Features are included for selecting different 'cakes' of data and automating fitting over many spectra, to enable tracking of peaks as they shift throughout the experiment. By analysing how different lattice plane peaks change during simulated processing, as can be seen in \autoref{fig:figure2}, the transformation and micromechanical behaviour of the material can be understood.
 
-![A schematic representing the features of ``xrdfit`` and the analysis of a synchrotron X-ray diffraction (SXRD) pattern, showing: (a) a polar plot of the caked intensity data; (b) an option for selecting different cakes and merging caked datasets; (c) adjustable peak bounds (grey) and adjustable maxima and maxima bounds (red and green) for constraining the peak fit; (d) an example fit of multiple and overlapping peaks.\label{fig:figure1}](figure1.eps)
+![A schematic representing the features of ``xrdfit`` and the analysis of a synchrotron X-ray diffraction (SXRD) pattern, showing: (a) a polar plot of the caked intensity data; (b) an option for selecting different cakes and merging caked datasets; (c) adjustable peak bounds (grey) and adjustable maxima and minima bounds (red and green) for constraining the peak fit; (d) an example fit of multiple and overlapping peaks.\label{fig:figure1}](figure1.eps)
 
-``xrdfit`` is designed to be accessible for all researchers who need to process SXRD (and XRD) spectra and so does not require a detailed knowledge of programming or fitting. The package has been used for the analysis of data taken during the hot deformation of a two-phase titanium alloy, which is presented in an article currently in press [@Daniel_2019], and will be used for future studies investigating the high temperature processing of metals in our research group. We hope that its public release will allow other researchers to benefit from fast data fitting, reducing the effort required to analyse their experimental data.
+``xrdfit`` is designed to be accessible for all researchers who need to process SXRD (and XRD) spectra and so does not require a detailed knowledge of programming or fitting. The package has been used for the analysis of data taken during the hot deformation of a two-phase titanium alloy, which is presented in an article currently in press [@Daniel_2019], and will be used for future studies investigating the high temperature processing of metals in our research group. We hope that its public release will allow other researchers to benefit from fast data fitting (up to 40 times faster than the current software), reducing the effort required to analyse their experimental data.
 
 ![An example analysis of a two-phase titanium (Ti-6Al-4V) alloy during high temperature and high strain rate deformation, showing characteristic shifts of the $\alpha$ (0002) peak centre. The shifts of the peak can be used to calculate elastic lattice straining in the hexagonal close-packed (hcp) lattice, as well as measure the thermal contraction on cooling.\label{fig:figure2}](figure2.eps)
 
