### 2020-08-19 #2381 @merrygoat
@jgostick Thank you.

- ORCIDs and Names checked
- Release v1.1.1 - https://github.com/LightForm-group/xrdfit/releases/tag/v1.1.1
- Zenodo DOI: 10.5281/zenodo.3991638
- Names match between Zenodo and paper
