### 2020-08-24 #2381 @mikapfl
In the generated pdf (https://github.com/openjournals/joss-papers/blob/5c52f575ebfaac4c6c44846620f181dd6859ae5a/joss.02381/10.21105.joss.02381.pdf) the citations at the end of the first paragraph look funny. They somehow use `[]` instead of `()` and the `Hammersley_2016` also doesn't look right.
