### 2020-08-24 5863cc8
Edit to paper.md and paper.bib
Correct spelling of 'The University of Manchester' in paper.md
Edit Hammersley_2016 in paper.bib to replace quotes with brackets
