+++ paper/paper.bib
@@ -150,18 +150,18 @@
 }
 
 @article{Hammersley_2016,
-author = "Hammersley, A. P.",
-title = "{{\it FIT2D}: a multi-purpose data reduction, analysis and visualization program}",
-journal = "Journal of Applied Crystallography",
-year = "2016",
-volume = "49",
-number = "2",
-pages = "646--652",
-month = "Apr",
-doi = {10.1107/S1600576716000455},
-url = {https://doi.org/10.1107/S1600576716000455},
-abstract = {{\it FIT2D}~is one of the principal area detector data reduction, analysis and visualization programs used at the European Synchrotron Radiation Facility and is also used by more than 400 research groups worldwide, including many other synchrotron radiation facilities. It has been developed for X-ray science, but is applicable to other structural techniques and is used in analysing electron diffraction data and microscopy, and neutron diffraction and scattering data. {\it FIT2D}~works for both interactive and `batch'-style data processing. Calibration and correction of detector distortions, integration of two-dimensional data to a variety of one-dimensional scans, and one- and two-dimensional model fitting are the main uses. Many other general-purpose image processing and image visualization operations are available. Commands are available through a `graphical user interface' and operations common to certain types of analysis are grouped within `interfaces'. Executable versions for most workstation and personal computer systems, and web page documentation, are available at http://www.esrf.eu/computing/scientific/FIT2D.},
-keywords = {data analysis, data display, data reduction, data visualization, detector calibration, image processing, model fitting, powder diffraction, small-angle scattering, wide-angle scattering},
+author     = {Hammersley, A. P.},
+title      = {{\it FIT2D}: a multi-purpose data reduction, analysis and visualization program},
+journal    = {Journal of Applied Crystallography},
+year       = {2016},
+volume     = {49},
+number     = {2},
+pages      = {646--652},
+month      = {Apr},
+doi        = {10.1107/S1600576716000455},
+url        = {https://doi.org/10.1107/S1600576716000455},
+abstract   = {{\it FIT2D}~is one of the principal area detector data reduction, analysis and visualization programs used at the European Synchrotron Radiation Facility and is also used by more than 400 research groups worldwide, including many other synchrotron radiation facilities. It has been developed for X-ray science, but is applicable to other structural techniques and is used in analysing electron diffraction data and microscopy, and neutron diffraction and scattering data. {\it FIT2D}~works for both interactive and `batch'-style data processing. Calibration and correction of detector distortions, integration of two-dimensional data to a variety of one-dimensional scans, and one- and two-dimensional model fitting are the main uses. Many other general-purpose image processing and image visualization operations are available. Commands are available through a `graphical user interface' and operations common to certain types of analysis are grouped within `interfaces'. Executable versions for most workstation and personal computer systems, and web page documentation, are available at http://www.esrf.eu/computing/scientific/FIT2D.},
+keywords   = {data analysis, data display, data reduction, data visualization, detector calibration, image processing, model fitting, powder diffraction, small-angle scattering, wide-angle scattering},
 }
 
 @article{Rietveld_1969,
