+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-The evolution of peak profiles in synchrotron X-ray diffraction (SXRD) data can tell us how the internal crystallographic structures of metals change during applied heating, high temperature straining and cooling cycles [@Daniel_2019; @Stark_2015; @Canelo_Yubero_2016; @Hu_2017], which is invaluable information used to improve industrial processing routes [@Salem_2008]. The experiment requires a beamline, at a synchrotron radiation facility such as Diamond Light Source [@Diamond_2020], to produce a high energy X-ray beam and illuminate a polycrystalline sample [@Daniel_2019]. The results are recorded in the form of time-resolved diffraction pattern rings, which are converted into a spectra of intensity peaks versus two-theta angle for a given direction [@Filik_2017; @Ashiotis_2015; @Hammersley_1996]. However, since many intensity profiles are collected during each experiment, with detectors recording at speeds greater than 250 Hz [@PILATUS_2020; @Loeliger_2012], fitting each of the individual lattice plane peaks can take a long time using current available software [@Basham_2015; @Merkel_2015; Hammersley_2016].
+The evolution of peak profiles in synchrotron X-ray diffraction (SXRD) data can tell us how the internal crystallographic structures of metals change during applied heating, high temperature straining and cooling cycles [@Daniel_2019; @Stark_2015; @Canelo_Yubero_2016; @Hu_2017], which is invaluable information used to improve industrial processing routes [@Salem_2008]. The experiment requires a beamline, at a synchrotron radiation facility such as Diamond Light Source [@Diamond_2020], to produce a high energy X-ray beam and illuminate a polycrystalline sample [@Daniel_2019]. The results are recorded in the form of time-resolved diffraction pattern rings, which are converted into a spectra of intensity peaks versus two-theta angle for a given direction [@Filik_2017; @Ashiotis_2015; @Hammersley_1996]. However, since many intensity profiles are collected during each experiment, with detectors recording at speeds greater than 250 Hz [@PILATUS_2020; @Loeliger_2012], fitting each of the individual lattice plane peaks can take a long time using current available software [@Basham_2015; @Merkel_2015; @Hammersley_2016].
 
 There are existing packages which can be used to fit peaks in SXRD spectra, examples include DAWN [@Basham_2015], Multifit/Polydefix [@Merkel_2015] and Fit2d [@Hammersley_2016]. In these cases, the software are compiled packages with a graphical user interface. Setting up the peak fits usually involves a point and click method to select the peak bounds, meaning it is unlikely to create a repeatable analysis. These packages also struggle to distinguish any peaks that overlap, which is important for capturing changes in multi-phase materials [@Daniel_2019]. MAUD [@Lutterotti_2014] is a software package that approaches fitting in a different way. MAUD uses the Rietveld refinement method [@Rietveld_1969] to match a model of the beamline setup and material properties to the data. This method allows determination of additional material properties, such as crystallographic texture, but applies an averaging over the peak positions and intensities to fit the model, meaning individual peak shifts cannot be accurately determined.
 
