### 2020-08-24 #2381 @merrygoat
@mikapfl Thank you - you are right, we had missed an @ symbol there. Corrected and released as version v1.1.2.

@jgostick Is it possible to tell whedon to use another version? The updated doi is: `10.5281/zenodo.3997251`

We would also like to thank both reviewers - your comments were insightful and lead to significant improvements.
