### 2020-08-26 #2381 @Kevin-Mattheus-Moerman
@merrygoat These are some required steps before we proceed:

- [ ] Please amend the archive meta data, in terms of the title and authors (you can edit this on ZENODO) such that it matches your paper (@jgostick, thanks for handling this submission. For future reference this :point_up: is a check the handling editor should also).
- [ ] I've checked your paper and it seems in order. This work is about to be processed for acceptance. If you haven't already re-read your paper now is the time to give it a final check (in particular author names, affiliations, and acknowledgements).
